requirements.txt
alembic==1.8.1
blinker==1.5
click==8.1.3
Flask==2.2.2
flask-form==1.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2==2.9.5
python-dotenv==0.21.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
WTForms==3.0.1
